// src/typings/user.d.ts
export type MemberMemberAreaType = {
  /** 地区值 */
  codeVal: string
  /** 国家名称 */
  codeKey: string
  /** 是否可用  */
  enableInd?: number
  /** 国家缩写 */
  remark: string
}